move dependencies test to new file
