+++ test/index.spec.ts
@@ -1,7 +1,4 @@
 import {describe, expect, test} from 'vitest';
-import {readFile} from 'node:fs/promises';
-import {join, resolve} from 'node:path';
-import resolvePackagePath from 'resolve-package-path';
 import {
   cellToBoundary,
   tileToCell,
@@ -35,31 +32,6 @@ const TEST_TILES = [
   {x: 1, y: 2, z: 3, q: 5202361257054699519n},
   {x: 1023, y: 2412, z: 23, q: 5291729562728627583n}
 ];
-
-test('dependencies support esm', async () => {
-  // Avoid CJS-only dependencies, which cause issues for downstream users.
-  // TODO: After Node.js v22 is the lowest version used in CI, replace the
-  // `new URL(...)` and `resolve-package-path` lookups with the built-in
-  // `module.findPackageJsonPath` in Node.js v22+.
-  const pkgQueue = [new URL('../package.json', import.meta.url).pathname];
-  const pkgVisited = new Set<string>();
-
-  for (const path of pkgQueue) {
-    const pkg = JSON.parse(await readFile(path, 'utf8'));
-
-    expect(pkg.type, `${pkg.name} must be esm`).toBe('module');
-
-    pkgVisited.add(pkg.name);
-
-    const dependencies = pkg.dependencies ? Object.keys(pkg.dependencies) : [];
-    const peerDependencies = pkg.peerDependencies ? Object.keys(pkg.peerDependencies) : [];
-    for (const dep of [...dependencies, ...peerDependencies]) {
-      if (!pkgVisited.has(dep)) {
-        pkgQueue.push(resolvePackagePath(dep, path)!);
-      }
-    }
-  }
-});
 
 describe('tileToCell', () => {
   test.each(TEST_TILES)('%s', ({x, y, z, q}) => {
